+++ src/engine_core/app_loader/src/ApplicationLoader.hpp
@@ -8,7 +8,6 @@
 
 #include "IApplication.hpp"
 #include <memory>
-#include "AppSupport.hpp"
 
 namespace Hush
 {
@@ -21,5 +20,5 @@ namespace Hush
 	/// If a static application is bundled with the engine, it won't attempt to load a shared library.
 	///
 	/// @return A pointer to the loaded application.
-	// std::unique_ptr<IApplication> LoadApplication(HushEngine *engine);
+	std::unique_ptr<IApplication> LoadApplication(HushEngine *engine);
 } // namespace Hush
